+++ tests/test_voc_builder.py
@@ -4,11 +4,12 @@ import pytest
 
 from voc_builder.main import (
     VocBuilderCSVFile,
-    WordSample,
+    WordInvalidForAdding,
     get_mastered_word_store,
-    parse_openai_reply,
+    validate_result_word,
     write_new_one,
 )
+from voc_builder.models import WordSample
 
 
 @pytest.fixture(autouse=True)
@@ -68,64 +69,39 @@ class TestVocBuilderCSVFile:
 
         # re-initialize the obj and append another word
         builder_2 = VocBuilderCSVFile(p)
-        builder_2.append_word(WordSample('Hello', '你好', 'həˈlō', 'Hello.', '你好。'))
+        builder_2.append_word(WordSample('hello', '你好', 'həˈlō', 'Hello.', '你好。'))
 
         words_2 = builder_2.read_all()
         assert len(words_2) == 2
-        assert words_2[1].word == 'Hello'
+        assert words_2[1].word == 'hello'
         assert words_2[1].word_meaning == '你好'
         assert words_2[1].translated_text == '你好。'
 
+    def test_remove_word(self, tmp_path, w_sample_world):
+        builder = VocBuilderCSVFile(tmp_path / 'foo.csv')
+        builder.append_word(w_sample_world)
+        builder.append_word(WordSample('hello', '你好', 'həˈlō', 'Hello.', '你好。'))
+        assert len(builder.read_all()) == 2
 
-@pytest.mark.parametrize(
-    'orig_text,input,expected',
-    [
-        # A standard reply
-        (
-            'Hello, world.',
-            'word: world\npronunciation: wɔːld\nmeaning: 世界\ntranslated: 你好，世界。',
-            WordSample(
-                word='world',
-                pronunciation='wɔːld',
-                word_meaning='世界',
-                orig_text='Hello, world.',
-                translated_text='你好，世界。',
-            ),
-        ),
-        # A reply using non-standard key name
-        (
-            'Hello, world.',
-            'Uncommon word: world\npronunciation: wɔːld\nmeaning: 世界\ntranslated: 你好，世界。',
-            WordSample(
-                word='world',
-                pronunciation='wɔːld',
-                word_meaning='世界',
-                orig_text='Hello, world.',
-                translated_text='你好，世界。',
-            ),
-        ),
-        # A reply which has extra content in the beginning and the end
-        (
-            'Hello, world.',
-            'foobar\n\nword: world\npronunciation: wɔːld\nmeaning: 世界\ntranslated: 你好，世界。\nother info',
-            WordSample(
-                word='world',
-                pronunciation='wɔːld',
-                word_meaning='世界',
-                orig_text='Hello, world.',
-                translated_text='你好，世界。',
-            ),
-        ),
-        (
-            'Hello, world.',
-            'invalid reply',
-            None,
-        ),
-    ],
-)
-def test_parse_openai_reply(orig_text, input, expected):
-    if expected is None:
-        with pytest.raises(ValueError):
-            parse_openai_reply(input, orig_text)
-    else:
-        assert parse_openai_reply(input, orig_text) == expected
+        builder.remove_words({'world'})
+        words = builder.read_all()
+        assert len(words) == 1
+        assert words[0].word == 'hello'
+
+
+def test_validate_result_word_misc(tmp_path):
+    builder = VocBuilderCSVFile(tmp_path / 'foo.csv')
+    validate_result_word(WordSample.make_empty('foo'), 'foo bar', builder)
+
+    with pytest.raises(WordInvalidForAdding, match='already in your vocabulary book'):
+        builder.append_word(WordSample.make_empty('foo'))
+        validate_result_word(WordSample.make_empty('foo'), 'foo bar', builder)
+    builder.remove_words({'foo'})
+
+    with pytest.raises(WordInvalidForAdding, match='already mastered'):
+        get_mastered_word_store().add('foo')
+        validate_result_word(WordSample.make_empty('foo'), 'foo bar', builder)
+    get_mastered_word_store().remove('foo')
+
+    with pytest.raises(WordInvalidForAdding, match='not in the original text'):
+        validate_result_word(WordSample.make_empty('foo'), 'bar baz', builder)
