+++ voc_builder/interactive.py
@@ -1,0 +1,7 @@
+# Special commands
+# discard last added word, try to get other options and let user choose manually
+COMMAND_NO = 'no'
+
+
+def handle_command_no():
+    pass
